# chat-backend/swagger.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  description: |
    Backend API for RAG-based Financial Products Chatbot with document upload, processing, and retrieval evaluation capabilities.

    ## Features
    - Multi-model LLM support (Gemma3, Qwen, Phi3, DeepSeek, SmolLM)
    - Dynamic document upload and processing
    - ChromaDB collection management
    - Retrieval evaluation metrics (MRR, Recall@K, Precision@K, NDCG@K)
    - Real-time chat with source citations
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:5000
    description: Local development server

tags:
  - name: Chat
    description: Chat and question-answering endpoints
  - name: Documents
    description: Document upload and processing
  - name: Collections
    description: ChromaDB collection management
  - name: Evaluation
    description: Retrieval quality evaluation
  - name: System
    description: System settings and health checks

paths:
  /api/chat:
    post:
      tags:
        - Chat
      summary: Chat with the RAG system
      description: Submit a question and receive an AI-generated answer with source citations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: User's question
                  example: "What is Bancassurance?"
                model:
                  type: string
                  description: LLM model to use
                  default: "gemma3:1b"
                  enum:
                    - gemma3:1b
                    - qwen:latest
                    - Phi3:latest
                    - deepseek-r1:1.5b
                    - smollm:latest
                temperature:
                  type: number
                  format: float
                  description: Sampling temperature (0.0-1.0)
                  default: 0.7
                  minimum: 0.0
                  maximum: 1.0
                top_k:
                  type: integer
                  description: Number of documents to retrieve
                  default: 3
                  minimum: 1
                  maximum: 20
                collection_name:
                  type: string
                  description: ChromaDB collection to query (default is 'data')
                  default: "data"
                  example: "uploaded_abc123"
      responses:
        "200":
          description: Successful response with answer and sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  answer:
                    type: string
                    example: "Bancassurance is a systematic marketing, distribution and servicing of a range of insurance products..."
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "Bancassurance Presentation::0"
                        source:
                          type: string
                          example: "Bancassurance Presentation.json"
                        text:
                          type: string
                          description: First 500 characters of the source text
                  model:
                    type: string
                    example: "gemma3:1b"
                  collection:
                    type: string
                    example: "data"
                  latency:
                    type: number
                    format: float
                    example: 1.23
        "400":
          description: Bad request (missing question or invalid model)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/upload:
    post:
      tags:
        - Documents
      summary: Upload documents
      description: Upload PDF and JSON files for processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: PDF or JSON files (max 50MB each)
      responses:
        "200":
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  session_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                          example: "loan_policy.pdf"
                        path:
                          type: string
                          example: "/uploads/123e4567.../loan_policy.pdf"
                        type:
                          type: string
                          example: "pdf"
                        size:
                          type: integer
                          example: 2048000
                  errors:
                    type: array
                    items:
                      type: string
                    nullable: true
        "400":
          description: No files provided or invalid files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/process/{session_id}:
    post:
      tags:
        - Documents
      summary: Process uploaded documents
      description: Extract text, chunk, and ingest documents into ChromaDB
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Upload session ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                collection_name:
                  type: string
                  description: Name for the ChromaDB collection
                  example: "my_banking_docs"
                chunk_size:
                  type: integer
                  description: Size of text chunks in characters
                  default: 1000
                  example: 1000
                overlap:
                  type: integer
                  description: Overlap between chunks in characters
                  default: 200
                  example: 200
      responses:
        "200":
          description: Documents processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  collection_name:
                    type: string
                    example: "uploaded_abc123"
                  num_documents:
                    type: integer
                    example: 3
                  num_chunks:
                    type: integer
                    example: 152
                  documents:
                    type: array
                    items:
                      type: object
                      properties:
                        doc_id:
                          type: string
                        num_chunks:
                          type: integer
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Processing error (no PDFs found, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/sessions/{session_id}/files:
    get:
      tags:
        - Documents
      summary: List session files
      description: Get list of all files in an upload session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  session_id:
                    type: string
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                        path:
                          type: string
                        type:
                          type: string
                        size:
                          type: integer

  /api/sessions/{session_id}:
    delete:
      tags:
        - Documents
      summary: Delete upload session
      description: Delete session and all uploaded files
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string

  /api/collections:
    get:
      tags:
        - Collections
      summary: List all collections
      description: Get list of all ChromaDB collections
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  collections:
                    type: array
                    items:
                      type: string
                    example: ["data", "uploaded_abc123", "my_banking_docs"]

  /api/collections/{collection_name}:
    delete:
      tags:
        - Collections
      summary: Delete a collection
      description: Delete a ChromaDB collection
      parameters:
        - name: collection_name
          in: path
          required: true
          schema:
            type: string
          description: Name of collection to delete
      responses:
        "200":
          description: Collection deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
        "500":
          description: Failed to delete collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/evaluate/retrieval:
    post:
      tags:
        - Evaluation
      summary: Evaluate retrieval quality
      description: |
        Evaluate retrieval performance using standard IR metrics:
        - MRR (Mean Reciprocal Rank)
        - Recall@K
        - Precision@K
        - NDCG@K (Normalized Discounted Cumulative Gain)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - test_queries
              properties:
                collection_name:
                  type: string
                  description: Collection to evaluate
                  default: "data"
                test_queries:
                  type: array
                  description: Test queries with ground truth relevance
                  items:
                    type: object
                    required:
                      - query
                      - relevant_doc_ids
                    properties:
                      query:
                        type: string
                        example: "What is Bancassurance?"
                      relevant_doc_ids:
                        type: array
                        items:
                          type: string
                        example:
                          [
                            "Bancassurance Presentation::0",
                            "Bancassurance Presentation::1",
                          ]
                top_k:
                  type: integer
                  description: Number of documents to retrieve for evaluation
                  default: 5
                  minimum: 1
                  maximum: 20
            example:
              collection_name: "data"
              test_queries:
                - query: "What is Bancassurance?"
                  relevant_doc_ids: ["Bancassurance Presentation::0"]
                - query: "What are HBL car loan rates?"
                  relevant_doc_ids: ["HBL CarLoan::2", "HBL CarLoan::3"]
              top_k: 5
      responses:
        "200":
          description: Evaluation results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  collection:
                    type: string
                    example: "data"
                  num_queries:
                    type: integer
                    example: 2
                  top_k:
                    type: integer
                    example: 5
                  mrr:
                    type: number
                    format: float
                    description: Mean Reciprocal Rank
                    example: 0.75
                  recall_at_k:
                    type: number
                    format: float
                    description: Recall at K
                    example: 0.85
                  precision_at_k:
                    type: number
                    format: float
                    description: Precision at K
                    example: 0.60
                  ndcg_at_k:
                    type: number
                    format: float
                    description: Normalized Discounted Cumulative Gain at K
                    example: 0.82
                  details:
                    type: array
                    description: Per-query metrics
                    items:
                      type: object
                      properties:
                        query:
                          type: string
                        mrr:
                          type: number
                        recall_at_k:
                          type: number
                        precision_at_k:
                          type: number
                        ndcg_at_k:
                          type: number
        "400":
          description: Missing or invalid test queries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/models:
    get:
      tags:
        - System
      summary: Get available models
      description: List all available LLM models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: string
                    example:
                      [
                        "gemma3:1b",
                        "qwen:latest",
                        "Phi3:latest",
                        "deepseek-r1:1.5b",
                        "smollm:latest",
                      ]

  /api/settings:
    get:
      tags:
        - System
      summary: Get default settings
      description: Get default configuration values
      responses:
        "200":
          description: Default settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  defaultModel:
                    type: string
                    example: "gemma3:1b"
                  temperature:
                    type: number
                    example: 0.7
                  topK:
                    type: integer
                    example: 3
                  models:
                    type: array
                    items:
                      type: string

  /api/health:
    get:
      tags:
        - System
      summary: Health check
      description: Check if the API is running
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /api/debug/chroma:
    post:
      tags:
        - System
      summary: Debug ChromaDB
      description: Test ChromaDB retrieval (debug endpoint)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  default: "What is HBL?"
                top_k:
                  type: integer
                  default: 3
      responses:
        "200":
          description: Debug results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  question:
                    type: string
                  num_results:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object

components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        error:
          type: string
          example: "Error message describing what went wrong"

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not currently implemented)
# Security can be added later
# security:
#   - ApiKeyAuth: []